Revert "Use webpack to build both dev and prod servers"
