+++ .eslintignore
@@ -6,3 +6,4 @@ static/*
 static-analysis/*
 dist/*
 node_modules/*
+webpack.config.js
