+++ docker-compose.yml
@@ -10,14 +10,12 @@ services:
       - 8000:8000
     depends_on:
       - platform
-      - pipeline
-      - search
   broker:
     image: rmohr/activemq
     ports:
       - 61613:61613
   platformdata:
-    image: ld4p/sinopia_acl:latest
+    image: ld4p/sinopia_acl:0.2
     environment:
       INSIDE_CONTAINER: 'true'
       TRELLIS_BASE_URL: http://platform:8080
@@ -44,36 +42,10 @@ services:
       - broker
       - database
       - migration
-  pipeline:
-    image: ld4p/sinopia_indexing_pipeline:latest
-    environment:
-      INDEX_URL: http://search:9200
-      BROKER_HOST: broker
-    depends_on:
-      - broker
-      - search
-      - platform
-  search:
-    image: docker.elastic.co/elasticsearch/elasticsearch:6.3.2
-    entrypoint:
-      - elasticsearch
-      - -Ehttp.port=9200
-      - -Ehttp.cors.enabled=true
-      - -Ehttp.cors.allow-origin=http://searchui:1358,http://localhost:1358,http://127.0.0.1:1358
-      - -Ehttp.cors.allow-headers=X-Requested-With,X-Auth-Token,Content-Type,Content-Length,Authorization
-      - -Ehttp.cors.allow-credentials=true
-      - -Etransport.host=localhost
-      - -Ebootstrap.system_call_filter=false
-    user: elasticsearch
+  broker:
+    image: rmohr/activemq
     ports:
-      - 9200:9200
-      - 9300:9300
-  searchui:
-    image: appbaseio/dejavu:latest
-    ports:
-      - 1358:1358
-    depends_on:
-      - search
+      - 61613:61613
   database:
     image: postgres:latest
     environment:
