+++ index.html
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
+
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
 
     <!-- Latest compiled and minified JavaScript -->
@@ -12,9 +13,11 @@
     <!-- Optional theme -->
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css" integrity="sha384-rHyoN1iRsVXV4nD0JutlnGaslCJuC7uwjduW9SVrLvRYooPp2bWYgmgJQIXwl/Sp" crossorigin="anonymous">
 
+
     <title>Sinopia Linked Data Editor</title>
   </head>
   <body>
     <div id="app"></div>
+    <script src="/dist/bundle.js"></script>
   </body>
 </html>
