+++ jest-puppeteer.config.js
@@ -7,6 +7,7 @@ module.exports = {
   },
   server: {
     command: 'SPOOF_SINOPIA_SERVER=true npm run dev-start',
+    port: 8888,
     launchTimeout: 10000,
   },
 }
