+++ package.json
@@ -36,8 +36,6 @@
     "babel-jest": "^24.7.1",
     "babel-loader": "^8.0.5",
     "babel-plugin-module-resolver": "^3.2.0",
-    "connect-history-api-fallback": "^1.6.0",
-    "copy-webpack-plugin": "^5.0.3",
     "coveralls": "^3.0.3",
     "css-loader": "^2.1.1",
     "enzyme": "^3.9.0",
@@ -54,7 +52,6 @@
     "eslint-plugin-security": "^1.4.0",
     "expect-puppeteer": "^4.1.1",
     "file-loader": "^3.0.1",
-    "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^24.7.1",
     "jest-junit": "^6.3.0",
@@ -63,18 +60,12 @@
     "jest-puppeteer": "^4.1.1",
     "jsdom": "15.0.0",
     "jsdom-global": "3.0.2",
-    "mini-css-extract-plugin": "^0.7.0",
-    "optimize-css-assets-webpack-plugin": "^5.0.1",
     "puppeteer": "^1.14.0",
     "react-hot-loader": "^4.8.4",
     "style-loader": "^0.23.1",
-    "terser-webpack-plugin": "^1.3.0",
-    "url-loader": "^2.0.0",
     "webpack": "^4.30.0",
     "webpack-cli": "^3.3.1",
-    "webpack-dev-middleware": "^3.7.0",
-    "webpack-hot-middleware": "^2.25.0",
-    "webpack-node-externals": "^1.7.2"
+    "webpack-dev-server": "^3.3.1"
   },
   "bugs": {
     "url": "http://github.com/LD4P/sinopia_editor/issues"
@@ -84,17 +75,16 @@
   },
   "scripts": {
     "analysis": "es6-plato -d static-analysis -n -e .eslintrc.js src src/lib",
-    "build": "webpack -p --progress --mode production --config webpack.prod.config.js && webpack -p --progress --mode production --config webpack.server.config.js",
+    "build": "webpack -p --progress --mode production",
     "ci": "npm i && npm run eslint && npm run dev-build && npm run jest-cov",
-    "dev-build": "webpack --progress --mode development --config webpack.dev.config.js && webpack --mode development --config webpack.server.config.js",
+    "dev-build": "webpack --progress --mode development",
     "dev-build-test": "npm run dev-build && npm run test",
-    "dev-start": "npm run dev-build && npm start",
-    "eslint": "eslint --max-warnings 11 --color -c .eslintrc.js --ext js,jsx .",
+    "dev-start": "webpack-dev-server --config ./webpack.config.js --mode development",
+    "eslint": "eslint --max-warnings 10 --color -c .eslintrc.js --ext js,jsx .",
     "eslint-fix": "eslint --max-warnings 10 --color -c .eslintrc.js --ext js,jsx --fix .",
     "eslint-fix-single": "eslint --color -c .eslintrc.js --fix",
     "jest-cov": "jest --coverage --colors --silent --runInBand",
     "jest-ci": "jest  --colors --silent --ci --runInBand --coverage --reporters=default --reporters=jest-junit && cat ./coverage/lcov.info | coveralls",
-    "start": "node ./dist/server.js",
     "test": "jest --colors --silent --runInBand --detectOpenHandles",
     "test-verbose": "jest --colors --runInBand"
   },
