+++ src/components/App.jsx
@@ -4,7 +4,7 @@ import React, { Component } from 'react'
 import 'react-bootstrap-typeahead/css/Typeahead.css'
 import LoginPanel from './LoginPanel'
 import HomePage from './home/HomePage'
-import '../css/main.css'
+import '../styles/main.css'
 import Editor from './editor/Editor'
 import Footer from './Footer'
 import { Route, Switch, withRouter } from 'react-router-dom'
