+++ src/styles/sinopia-logo.png
