+++ src/versoSpoof.js
@@ -14,7 +14,7 @@ function loadProfiles() {
     const profilesDirPath = path.join(__dirname, '..', 'static', 'spoofedFilesFromServer', 'from_verso', 'data', 'profiles')
 
     fs.readdirSync(profilesDirPath).forEach((file) => {
-      const fileJson = JSON.parse(fs.readFileSync(path.join(profilesDirPath, file)))
+      const fileJson = require(path.join(profilesDirPath, file))
 
       profiles.push(
         {
