+++ static/spoofedFilesFromServer/from_verso/data/profiles/BIBFRAME 2.0 DDC.json
