--- webpack.prod.config.js
@@ -1,91 +1,0 @@
-// Copyright 2018, 2019 Stanford University see LICENSE for license
-
-const path = require('path')
-const webpack = require('webpack')
-const HtmlWebpackPlugin = require('html-webpack-plugin')
-const CopyWebpackPlugin = require('copy-webpack-plugin')
-const MiniCssExtractPlugin = require('mini-css-extract-plugin')
-const TerserPlugin = require('terser-webpack-plugin')
-const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin')
-
-module.exports = {
-  entry: {
-    main: './src/index.js',
-  },
-  output: {
-    path: path.join(__dirname, '/dist'),
-    publicPath: '/',
-    filename: '[name].js',
-  },
-  target: 'web',
-  // devtool: 'source-map',
-  optimization: {
-    minimizer: [
-      new TerserPlugin({
-        cache: true,
-        parallel: true,
-        sourceMap: false, // set to true if you want JS source maps
-      }),
-      new OptimizeCSSAssetsPlugin({}),
-    ],
-  },
-  module: {
-    rules: [
-      {
-        test: /\.jsx?$/,
-        resolve: {
-          extensions: ['.js', '.jsx'],
-        },
-        exclude: /node_modules/,
-        use: ['babel-loader'],
-      },
-      {
-        test: /\.html$/,
-        use: [
-          {
-            loader: 'html-loader',
-            options: { minimize: true },
-          },
-        ],
-      },
-      {
-        test: /\.css$/,
-        use: [MiniCssExtractPlugin.loader, 'css-loader'],
-      },
-      {
-        test: /\.(png|jpg|gif)$/,
-        use: ['url-loader'],
-      },
-      {
-        parser: {
-          amd: false,
-        },
-      },
-    ],
-  },
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: './src/html/index.html',
-      filename: './index.html',
-    }),
-    new CopyWebpackPlugin([
-      { from: 'static', to: 'static' },
-    ]),
-    new MiniCssExtractPlugin({
-      filename: '[name].css',
-      chunkFilename: '[id].css',
-    }),
-    new webpack.EnvironmentPlugin(
-      [
-        'SPOOF_SINOPIA_SERVER',
-        'TRELLIS_BASE_URL',
-        'DEFAULT_PROFILE_SCHEMA_VERSION',
-        'SINOPIA_GROUP',
-        'SINOPIA_URI',
-        'AWS_COGNITO_DOMAIN',
-        'COGNITO_CLIENT_ID',
-        'COGNITO_USER_POOL_ID',
-      ],
-    ),
-  ],
-}
